fix: fixed issue with players getting stuck
Changelog:
Removed the wait time to buckle and unbuckle the seat belt
Removed the ProgressBar function and its Wait(3750) to make the action of buckle/unbuckle immediate.
Added the reset of exiting the vehicle when the seat belt is unbuckled
The exit control (Control 75) is now immediately re-enabled when the seat belt is unbuckled.
Improved UI handling
The NUI window is updated correctly when the seat belt is buckled and unbuckled without inconsistencies.
With this update, players can buckle and unbuckle without delays and without getting stuck in the vehicle.